app/requirements.txt
Django
djangorestframework
pillow
django-extensions
django-lifecycle